Ids as int

+++ WCFSample/Models/OrderItem.cs
@@ -6,9 +6,9 @@ namespace Models
     public class OrderItem
     {
         [DataMember]
-        public long Id { get; set; }
+        public int Id { get; set; }
         [DataMember]
-        public long OrderId { get; set; }
+        public int OrderId { get; set; }
         [DataMember]
         public int ProductId { get; set; }
         [DataMember]
